chore: cleanup, rename to nice filename
